settings: Corrected type for offset variables
Changed type of offset variables from off_t to size_t as expected
by called functions.

+++ subsys/settings/src/settings_fcb.c
@@ -159,7 +159,7 @@ static void settings_fcb_compress(struct settings_fcb *cf)
 			break;
 		}
 
-		off_t val1_off;
+		size_t val1_off;
 
 		rc = settings_line_name_read(name1, sizeof(name1), &val1_off,
 					     &loc1);
@@ -178,7 +178,7 @@ static void settings_fcb_compress(struct settings_fcb *cf)
 		copy = 1;
 
 		while (fcb_getnext(&cf->cf_fcb, &loc2.loc) == 0) {
-			off_t val2_off;
+			size_t val2_off;
 
 			rc = settings_line_name_read(name2, sizeof(name2),
 						     &val2_off, &loc2);
